refactor(page api): move to api controller
